extract_utils: Handle moved packages properly
When a package was moved to somewhere by using a spec like
'-app/test.apk:vendor/app/test.apk', the script tried to extract the file
from 'vendor/app/test.apk' rather than 'app/test.apk'. This is not the
expected behavior.

This patch introduced a new array PRODUCT_PACKAGES_TARGETS to store the
package list, and keep the PRODUCT_PACKAGES_LIST in a form like
'source:target' for the mapping between sources and targets.

Change-Id: I1e451c465b135a18a476eb870e38e90289c944b5